requirements.txt
django
gunicorn
django-heroku
requests
djangorestframework
psycopg2